+++ src/server/mod.rs
@@ -1113,6 +1113,7 @@ impl ServerState {
         account: Option<String>,
         filename: String,
         path: PathBuf,
+        scrubbed_override: Option<bool>,
     ) -> Option<JobRequest> {
         tracing::debug!("Pushing file upload id: {id}");
         let identifiers = Identifiers {
@@ -1130,6 +1131,7 @@ impl ServerState {
                     filename,
                     path: path.clone(),
                     cleanup_path: Some(path),
+                    scrubbed_override,
                 },
             },
         )
@@ -1452,6 +1454,7 @@ pub enum JobInput {
         filename: String,
         path: PathBuf,
         cleanup_path: Option<PathBuf>,
+        scrubbed_override: Option<bool>,
     },
     FromServiceLink {
         source: String,